Fixed the shaky hand icons bug
Shaky hand icons were meant to appear when to user selects a choice to
indicate that the choice is being processed. However, the icons were
not appearing at all. This was due to a bug in the code that
determines when to show the icons. The bug has been fixed and the
icons now appear correctly when a choice is selected.